Added global exception handling and swagger docs

+++ src/main/java/project/ForumWebApp/services/Implementations/CommentServiceImpl.java
@@ -3,12 +3,14 @@ package project.ForumWebApp.services.Implementations;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
 import project.ForumWebApp.config.AuthContextManager;
+import project.ForumWebApp.exceptions.AuthorizationException;
 import project.ForumWebApp.models.Comment;
 import project.ForumWebApp.models.Post;
 import project.ForumWebApp.models.DTOs.CommentCreateDTO;
@@ -17,49 +19,55 @@ import project.ForumWebApp.repository.CommentRepository;
 import project.ForumWebApp.repository.PostRepository;
 import project.ForumWebApp.services.CommentService;
 
+import static project.ForumWebApp.constants.ValidationConstants.*;
+
 @Service
 public class CommentServiceImpl implements CommentService {
-     
+
+
+
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
     private final ModelMapper modelMapper;
     private final AuthContextManager authContextManager;
 
     public CommentServiceImpl(CommentRepository commentRepository, ModelMapper modelMapper,
-                              PostRepository postRepository,AuthContextManager authContextManager) {
+                              PostRepository postRepository, AuthContextManager authContextManager) {
         this.commentRepository = commentRepository;
         this.modelMapper = modelMapper;
         this.postRepository = postRepository;
         this.authContextManager = authContextManager;
     }
 
+    @Override
     public List<CommentDTO> getCommentsByPostId(int postId) {
         List<Comment> comments = commentRepository.findByPostId(postId);
         return comments.stream()
-                       .map(comment -> modelMapper.map(comment, CommentDTO.class))
-                       .collect(Collectors.toList());
+                .map(comment -> modelMapper.map(comment, CommentDTO.class))
+                .collect(Collectors.toList());
     }
 
     @Override
     public List<CommentDTO> getAllComments() {
         List<Comment> comments = commentRepository.findAll();
         return comments.stream()
-                       .map(comment -> modelMapper.map(comment, CommentDTO.class))
-                       .collect(Collectors.toList());
+                .map(comment -> modelMapper.map(comment, CommentDTO.class))
+                .collect(Collectors.toList());
     }
 
     @Override
     public CommentDTO getCommentById(Integer id) {
         Comment comment = commentRepository.findById(id)
-                                           .orElseThrow(() -> new RuntimeException("Comment not found"));
+                .orElseThrow(() -> new EntityNotFoundException(COMMENT_WITH_PROVIDED_ID_DOES_NOT_EXIST));
         return modelMapper.map(comment, CommentDTO.class);
     }
 
     @Override
     @Transactional
     public CommentDTO createComment(int id, CommentCreateDTO commentDTO) {
+        Post post = postRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(POST_WITH_PROVIDED_ID_DOES_NOT_EXIST));
         Comment comment = modelMapper.map(commentDTO, Comment.class);
-        Post post  = postRepository.findById(id).get();
         comment.setPost(post);
         comment.setUser(authContextManager.getLoggedInUser());
         Comment savedComment = commentRepository.save(comment);
@@ -72,7 +80,12 @@ public class CommentServiceImpl implements CommentService {
     @Transactional
     public CommentDTO updateComment(Integer id, CommentDTO commentDTO) {
         Comment existingComment = commentRepository.findById(id)
-                                                   .orElseThrow(() -> new RuntimeException("Comment not found"));
+                .orElseThrow(() -> new EntityNotFoundException(COMMENT_WITH_PROVIDED_ID_DOES_NOT_EXIST));
+
+        if (!existingComment.getUser().getUsername().equals(authContextManager.getUsername())) {
+            throw new AuthorizationException(YOU_ARE_NOT_AUTHORIZED_TO_UPDATE_THIS_COMMENT);
+        }
+
         existingComment.setContent(commentDTO.getContent());
         Comment updatedComment = commentRepository.save(existingComment);
         return modelMapper.map(updatedComment, CommentDTO.class);
@@ -80,14 +93,24 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public void deleteComment(Integer id) {
+        Comment comment = commentRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(COMMENT_WITH_PROVIDED_ID_DOES_NOT_EXIST));
+
+        if (!comment.getUser().getUsername().equals(authContextManager.getUsername())) {
+            throw new AuthorizationException(YOU_ARE_NOT_AUTHORIZED_TO_UPDATE_THIS_COMMENT);
+        }
+
         commentRepository.deleteById(id);
     }
-
 
     @Override
     public boolean isOwner(int id) {
         String currentUsername = authContextManager.getUsername();
-        Comment comment = commentRepository.findById(id).orElse(null);
-        return comment != null && comment.getUser().getUsername().equals(currentUsername);
+        Comment comment = commentRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(COMMENT_WITH_PROVIDED_ID_DOES_NOT_EXIST));
+        if (!comment.getUser().getUsername().equals(currentUsername)) {
+            throw new AuthorizationException(YOU_ARE_NOT_AUTHORIZED_TO_UPDATE_THIS_COMMENT);
+        }
+        return true;
     }
 }
